requirements.txt
easyocr
requests
python-dotenv
fastapi
uvicorn[standard]
python-multipart
pytest
pytest-asyncio
ruff
